redo how we get dependencies installed
build-dep for Iceweasel in Debian Wheezy appears to be busted.  mach
bootstrap installs individual packages instead of using build-dep and is
more likely to be correct.  Using mach bootstrap also means that we
avoid any differences between the source package names we need to use
(iceweasel on Debian, firefox on Ubuntu, etc.).